fix(fsutil): isolate umask test under !windows build tag
